Refactor: Move semantic analyzer pass 3 to a new module
Just move things around, tweak inter-module references and add
necessary imports.

The third pass still directly accesses the second pass so there is
more work to do.
