+++ mypy_self_check.ini
@@ -11,3 +11,6 @@ warn_unused_ignores = True
 # historical exception
 [mypy-mypy.semanal]
 strict_optional = False
+
+[mypy-mypy.semanal_pass3]
+strict_optional = False
